Made a split of consumer and producer from queueService.
