+++ examples/Examples.AdvancedConfiguration/MessageHandlers/CustomNonCyclicMessageHandler.cs
@@ -1,5 +1,3 @@
-using RabbitMQ.Client.Core.DependencyInjection.MessageHandlers;
-using RabbitMQ.Client.Core.DependencyInjection.Services;
 using RabbitMQ.Client.Events;
 
 namespace Examples.AdvancedConfiguration.MessageHandlers
