+++ examples/Examples.ConsumerConsole/CustomAsyncNonCyclicMessageHandler.cs
@@ -1,8 +1,6 @@
 ﻿using Microsoft.Extensions.Logging;
 using System.Threading.Tasks;
 using RabbitMQ.Client.Core.DependencyInjection;
-using RabbitMQ.Client.Core.DependencyInjection.MessageHandlers;
-using RabbitMQ.Client.Core.DependencyInjection.Services;
 using RabbitMQ.Client.Events;
 
 namespace Examples.ConsumerConsole
