+++ examples/Examples.ConsumerConsole/Program.cs
@@ -2,7 +2,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using RabbitMQ.Client.Core.DependencyInjection;
 using System.IO;
-using RabbitMQ.Client.Core.DependencyInjection.Services;
 
 namespace Examples.ConsumerConsole
 {
