+++ src/RabbitMQ.Client.Core.DependencyInjection/Filters/IMessageHandlingFilter.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Threading.Tasks;
-using RabbitMQ.Client.Core.DependencyInjection.Services;
+using RabbitMQ.Client.Core.DependencyInjection.Services.Interfaces;
 using RabbitMQ.Client.Events;
 
 namespace RabbitMQ.Client.Core.DependencyInjection.Filters
@@ -15,6 +15,6 @@ namespace RabbitMQ.Client.Core.DependencyInjection.Filters
         /// </summary>
         /// <param name="next">Next action.</param>
         /// <returns>Function that could process received message.</returns>
-        Func<BasicDeliverEventArgs, IQueueService, Task> Execute(Func<BasicDeliverEventArgs, IQueueService, Task> next);
+        Func<BasicDeliverEventArgs, IConsumingService, Task> Execute(Func<BasicDeliverEventArgs, IConsumingService, Task> next);
     }
 }
