+++ src/RabbitMQ.Client.Core.DependencyInjection/Models/RabbitMqConnectionOptionsContainer.cs
@@ -1,19 +1,17 @@
 using System;
 using RabbitMQ.Client.Core.DependencyInjection.Configuration;
-using RabbitMQ.Client.Core.DependencyInjection.Services;
 
 namespace RabbitMQ.Client.Core.DependencyInjection.Models
 {
     /// <summary>
     /// A model that contains a link between an instance of message queuing service and its configuration.
-    /// A queuing service can be of type <see cref="IQueueService"/>, <see cref="IConsumingService"/> or <see cref="IProducingService"/>.
     /// </summary>
     public class RabbitMqConnectionOptionsContainer
     {
         /// <summary>
-        /// Queuing service identifier.
+        /// Service type.
         /// </summary>
-        public Guid Guid { get; set; }
+        public Type Type { get; set; }
 
         /// <summary>
         /// Queuing service options container.
