--- src/RabbitMQ.Client.Core.DependencyInjection/Services/IQueueService.cs
@@ -1,9 +1,0 @@
-﻿namespace RabbitMQ.Client.Core.DependencyInjection.Services
-{
-    /// <summary>
-    /// Custom RabbitMQ queue service interface.
-    /// </summary>
-    public interface IQueueService : IConsumingService, IProducingService
-    {
-    }
-}
