+++ src/RabbitMQ.Client.Core.DependencyInjection/Services/Interfaces/IMessageHandlingPipelineExecutingService.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using RabbitMQ.Client.Events;
 
-namespace RabbitMQ.Client.Core.DependencyInjection.Services
+namespace RabbitMQ.Client.Core.DependencyInjection.Services.Interfaces
 {
     /// <summary>
     /// Service that is responsible for handling received message in the pipeline of filters.
@@ -12,7 +12,7 @@ namespace RabbitMQ.Client.Core.DependencyInjection.Services
         /// Execute message handling pipeline.
         /// </summary>
         /// <param name="eventArgs">Received message.</param>
-        /// <param name="queueService">Queuing service.</param>
-        Task Execute(BasicDeliverEventArgs eventArgs, IQueueService queueService);
+        /// <param name="consumingService">Consuming service.</param>
+        Task Execute(BasicDeliverEventArgs eventArgs, IConsumingService consumingService);
     }
 }
