+++ src/RabbitMQ.Client.Core.DependencyInjection/Services/MessageHandlerContainerBuilder.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using RabbitMQ.Client.Core.DependencyInjection.InternalExtensions;
 using RabbitMQ.Client.Core.DependencyInjection.MessageHandlers;
 using RabbitMQ.Client.Core.DependencyInjection.Models;
+using RabbitMQ.Client.Core.DependencyInjection.Services.Interfaces;
 
 namespace RabbitMQ.Client.Core.DependencyInjection.Services
 {
@@ -18,23 +19,17 @@ namespace RabbitMQ.Client.Core.DependencyInjection.Services
         readonly IEnumerable<MessageHandlerOrderingModel> _orderingModels;
         readonly IEnumerable<IMessageHandler> _messageHandlers;
         readonly IEnumerable<IAsyncMessageHandler> _asyncMessageHandlers;
-        readonly IEnumerable<INonCyclicMessageHandler> _nonCyclicHandlers;
-        readonly IEnumerable<IAsyncNonCyclicMessageHandler> _asyncNonCyclicHandlers;
 
         public MessageHandlerContainerBuilder(
             IEnumerable<MessageHandlerRouter> routers,
             IEnumerable<MessageHandlerOrderingModel> orderingModels,
             IEnumerable<IMessageHandler> messageHandlers,
-            IEnumerable<IAsyncMessageHandler> asyncMessageHandlers,
-            IEnumerable<INonCyclicMessageHandler> nonCyclicHandlers,
-            IEnumerable<IAsyncNonCyclicMessageHandler> asyncNonCyclicHandlers)
+            IEnumerable<IAsyncMessageHandler> asyncMessageHandlers)
         {
             _routers = routers;
             _orderingModels = orderingModels;
             _messageHandlers = messageHandlers;
             _asyncMessageHandlers = asyncMessageHandlers;
-            _nonCyclicHandlers = nonCyclicHandlers;
-            _asyncNonCyclicHandlers = asyncNonCyclicHandlers;
         }
 
         /// <summary>
@@ -61,26 +56,20 @@ namespace RabbitMQ.Client.Core.DependencyInjection.Services
             return containers;
         }
 
-        MessageHandlerContainer CreateContainer(string exchange, IEnumerable<MessageHandlerRouter> selectedRouters)
+        MessageHandlerContainer CreateContainer(string exchange, IList<MessageHandlerRouter> selectedRouters)
         {
             var routersDictionary = TransformMessageHandlerRoutersToDictionary(selectedRouters);
-            var boundMessageHandlers = _messageHandlers.Where(x => routersDictionary.Keys.Contains(x.GetType()));
-            var boundAsyncMessageHandlers = _asyncMessageHandlers.Where(x => routersDictionary.Keys.Contains(x.GetType()));
-            var boundNonCyclicMessageHandlers = _nonCyclicHandlers.Where(x => routersDictionary.Keys.Contains(x.GetType()));
-            var boundAsyncNonCyclicMessageHandlers = _asyncNonCyclicHandlers.Where(x => routersDictionary.Keys.Contains(x.GetType()));
+            var boundMessageHandlers = _messageHandlers.Where(x => routersDictionary.Keys.Contains(x.GetType())).ToList();
+            var boundAsyncMessageHandlers = _asyncMessageHandlers.Where(x => routersDictionary.Keys.Contains(x.GetType())).ToList();
             var routePatterns = selectedRouters.SelectMany(x => x.RoutePatterns).Distinct().ToList();
             var messageHandlers = TransformMessageHandlersCollectionsToDictionary(
                 boundMessageHandlers,
                 boundAsyncMessageHandlers,
-                boundNonCyclicMessageHandlers,
-                boundAsyncNonCyclicMessageHandlers,
                 routersDictionary);
             var orderingModels = GetMessageHandlerOrderingModels(
                 exchange,
                 boundMessageHandlers,
                 boundAsyncMessageHandlers,
-                boundNonCyclicMessageHandlers,
-                boundAsyncNonCyclicMessageHandlers,
                 _orderingModels);
             return new MessageHandlerContainer
             {
@@ -112,14 +101,10 @@ namespace RabbitMQ.Client.Core.DependencyInjection.Services
             string exchange,
             IEnumerable<IMessageHandler> messageHandlers,
             IEnumerable<IAsyncMessageHandler> asyncMessageHandlers,
-            IEnumerable<INonCyclicMessageHandler> nonCyclicMessageHandlers,
-            IEnumerable<IAsyncNonCyclicMessageHandler> asyncNonCyclicMessageHandlers,
             IEnumerable<MessageHandlerOrderingModel> orderingModels)
         {
             var messageHandlersCollection = new List<IBaseMessageHandler>(messageHandlers);
             messageHandlersCollection.AddRange(asyncMessageHandlers);
-            messageHandlersCollection.AddRange(nonCyclicMessageHandlers);
-            messageHandlersCollection.AddRange(asyncNonCyclicMessageHandlers);
 
             var messageHandlerTypes = messageHandlersCollection.Select(x => x.GetType()).ToList();
             return orderingModels.Where(x => messageHandlerTypes.Contains(x.MessageHandlerType) && x.Exchange == exchange);
@@ -128,17 +113,11 @@ namespace RabbitMQ.Client.Core.DependencyInjection.Services
         static IDictionary<string, IList<IBaseMessageHandler>> TransformMessageHandlersCollectionsToDictionary(
             IEnumerable<IMessageHandler> messageHandlers,
             IEnumerable<IAsyncMessageHandler> asyncMessageHandlers,
-            IEnumerable<INonCyclicMessageHandler> nonCyclicMessageHandlers,
-            IEnumerable<IAsyncNonCyclicMessageHandler> asyncNonCyclicMessageHandlers,
             IDictionary<Type, List<string>> routersDictionary)
         {
             var transformedMessageHandlers = TransformMessageHandlersCollectionToDictionary(messageHandlers, routersDictionary);
             var transformedAsyncMessageHandlers = TransformMessageHandlersCollectionToDictionary(asyncMessageHandlers, routersDictionary);
-            var transformedNonCyclicHandlers = TransformMessageHandlersCollectionToDictionary(nonCyclicMessageHandlers, routersDictionary);
-            var transformedAsyncNonCyclicHandlers = TransformMessageHandlersCollectionToDictionary(asyncNonCyclicMessageHandlers, routersDictionary);
-            return transformedMessageHandlers.UnionKeysAndValues(transformedAsyncMessageHandlers)
-                .UnionKeysAndValues(transformedNonCyclicHandlers)
-                .UnionKeysAndValues(transformedAsyncNonCyclicHandlers);
+            return transformedMessageHandlers.UnionKeysAndValues(transformedAsyncMessageHandlers);
         }
 
         static IDictionary<string, IList<IBaseMessageHandler>> TransformMessageHandlersCollectionToDictionary<T>(
