+++ src/RabbitMQ.Client.Core.DependencyInjection/Services/MessageHandlingPipelineExecutingService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using RabbitMQ.Client.Core.DependencyInjection.Filters;
+using RabbitMQ.Client.Core.DependencyInjection.Services.Interfaces;
 using RabbitMQ.Client.Events;
 
 namespace RabbitMQ.Client.Core.DependencyInjection.Services
@@ -26,43 +27,38 @@ namespace RabbitMQ.Client.Core.DependencyInjection.Services
             _exceptionFilters = exceptionFilters ?? Enumerable.Empty<IMessageHandlingExceptionFilter>();
         }
 
-        /// <summary>
-        /// Execute message handling pipeline.
-        /// </summary>
-        /// <param name="eventArgs">Received message.</param>
-        /// <param name="queueService">Queuing service.</param>
-        public async Task Execute(BasicDeliverEventArgs eventArgs, IQueueService queueService)
+        public async Task Execute(BasicDeliverEventArgs eventArgs, IConsumingService consumingService)
         {
             try
             {
-                await ExecutePipeline(eventArgs, queueService).ConfigureAwait(false);
+                await ExecutePipeline(eventArgs, consumingService).ConfigureAwait(false);
             }
             catch (Exception exception)
             {
-                await ExecuteFailurePipeline(exception, eventArgs, queueService).ConfigureAwait(false);
+                await ExecuteFailurePipeline(exception, eventArgs, consumingService).ConfigureAwait(false);
             }
         }
 
-        async Task ExecutePipeline(BasicDeliverEventArgs eventArgs, IQueueService queueService)
+        async Task ExecutePipeline(BasicDeliverEventArgs eventArgs, IConsumingService consumingService)
         {
-            Func<BasicDeliverEventArgs, IQueueService, Task> handle = _messageHandlingService.HandleMessageReceivingEvent;
+            Func<BasicDeliverEventArgs, IConsumingService, Task> handle = _messageHandlingService.HandleMessageReceivingEvent;
             foreach (var filter in _handlingFilters.Reverse())
             {
                 handle = filter.Execute(handle);
             }
 
-            await handle(eventArgs, queueService).ConfigureAwait(false);
+            await handle(eventArgs, consumingService).ConfigureAwait(false);
         }
 
-        async Task ExecuteFailurePipeline(Exception exception, BasicDeliverEventArgs eventArgs, IQueueService queueService)
+        async Task ExecuteFailurePipeline(Exception exception, BasicDeliverEventArgs eventArgs, IConsumingService consumingService)
         {
-            Func<Exception, BasicDeliverEventArgs, IQueueService, Task> handle = _messageHandlingService.HandleMessageProcessingFailure;
+            Func<Exception, BasicDeliverEventArgs, IConsumingService, Task> handle = _messageHandlingService.HandleMessageProcessingFailure;
             foreach (var filter in _exceptionFilters.Reverse())
             {
                 handle = filter.Execute(handle);
             }
 
-            await handle(exception, eventArgs, queueService).ConfigureAwait(false);
+            await handle(exception, eventArgs, consumingService).ConfigureAwait(false);
         }
     }
 }
