+++ src/RabbitMQ.Client.Core.DependencyInjection/Services/RabbitMqConnectionFactory.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading;
 using RabbitMQ.Client.Core.DependencyInjection.Configuration;
 using RabbitMQ.Client.Core.DependencyInjection.Exceptions;
+using RabbitMQ.Client.Core.DependencyInjection.Services.Interfaces;
 using RabbitMQ.Client.Events;
 using RabbitMQ.Client.Exceptions;
