+++ tests/RabbitMQ.Client.Core.DependencyInjection.Tests/Stubs/StubAsyncNonCyclicMessageHandler.cs
@@ -1,6 +1,4 @@
 using System.Threading.Tasks;
-using RabbitMQ.Client.Core.DependencyInjection.MessageHandlers;
-using RabbitMQ.Client.Core.DependencyInjection.Services;
 using RabbitMQ.Client.Events;
 
 namespace RabbitMQ.Client.Core.DependencyInjection.Tests.Stubs
